Bugfix: Improve status monitoring
Status monitoring now gets the current status id into the TogglTimerID which means that when the device turns on it sets that variables and can ensure it can turn off even if it wasn't started during this devices runtime.

This means that if the device power cycles (which it probably will because it's an ESP) then it will be able to pickup the running timer and crack on as if it started it.